read context from github repo and workspace
