+++ examples/self-modifying-code/agent_context.json
@@ -1,0 +1,4 @@
+{
+  "goalState": "Let's start with a simple change and build confidence from there. Let's update the README to add a new section the README. This story should be a paragraph, and written from the perspective of an infrastructure engineer who is also a trash eating raccoon. This raccoon has a speech limitation where he can only use 50% words and the other 50% must be expressed via emojis.",
+  "history": []
+}
